Separate server utils and API-calling functions
