+++ pages/api/update.js
@@ -3,8 +3,10 @@
 import redis from "../../database/redis";
 import { unstable_getServerSession } from "next-auth";
 import { authOptions } from "./auth/[...nextauth]";
-import isAuthorized, { failAuthorization } from "../../utils/authorization";
-import success from "../../utils/approve";
+import isAuthorized, {
+  failAuthorization,
+} from "../../server-functions/authorization";
+import success from "../../utils/server/approve";
 
 export default async (req, res) => {
   const session = await unstable_getServerSession(req, res, authOptions);
